- Removed deprecated `MapPolygonActivity`, `SocketManager`, and `WebSocketManager` to streamline the codebase and prioritize SignalR for real-time tracking. - Refactored `LocationService` to optimize performance by removing `WakeLock` and implementing a 5-second threshold and movement check (0.0001 degrees) for location updates. - Renamed `TransferReciveDetail` to `TransferReceiveDetail` and updated all related activities and adapters for consistency. - Standardized UI colors by renaming `darkgray` to `darkGray` across XML layouts and drawable resources. - Updated `DeliveryOrderDetailActivity` and `ConfirmDeliveryActivity` to use string resources for status labels like "Open," "InProgress," and "Delivered" to support localization. - Improved UI consistency by replacing `<fragment>` tags with `FragmentContainerView` in `activity_route_map.xml` and `activity_map_route.xml`. - Refactored `RequisitionListAdapter` and `PaymentLogsListAdapter` to use `when` expressions for cleaner status-based background styling. - Fixed several typos in layout files (e.g., "Kilomerers" to "Kilometers") and standardized method names like `openImageChooseDialog`. - Updated dependencies for `activity-ktx`, `camera`, and `signalr` to their latest versions in `app/build.gradle`. - Cleaned up `AlvimaTuckApplication` and `Utils.kt` by removing unused helper methods and optimizing bitmap generation using `createBitmap`.

+++ app/src/main/java/com/alvimatruck/activity/HomeActivity.kt
@@ -245,7 +245,7 @@ class HomeActivity : BaseActivity<ActivityHomeBinding>() {
                             binding.tvTodayVisit.text =
                                 dashboardDetails!!.totalCustomerVisits.toString()
                             binding.tvTotalSaleCash.text =
-                                "ETB " + dashboardDetails!!.todaySalesSum.toDouble().to2Decimal()
+                                "ETB " + dashboardDetails!!.todaySalesSum.to2Decimal()
                             binding.tvTotalCollectionCash.text =
                                 dashboardDetails!!.todayCollectionsCount.toString()
 
@@ -264,9 +264,7 @@ class HomeActivity : BaseActivity<ActivityHomeBinding>() {
 
                 override fun onFailure(call: Call<JsonObject?>, t: Throwable) {
                     Toast.makeText(
-                        this@HomeActivity,
-                        getString(R.string.api_fail_message),
-                        Toast.LENGTH_SHORT
+                        this@HomeActivity, getString(R.string.api_fail_message), Toast.LENGTH_SHORT
                     ).show()
                     ProgressDialog.dismiss()
                 }
